add icons in editor title to switch between drawing and source
