+++ extension/package.json
@@ -23,8 +23,8 @@
         "url": "https://github.com/pomdtr"
     },
     "activationEvents": [
-        "onCustomEditor:editor.excalidraw",
-        "onCommand:excalidraw.theme.update"
+        "workspaceContains:**/*.{excalidraw,excalidraw.svg}",
+        "onCustomEditor:editor.excalidraw"
     ],
     "repository": {
         "type": "git",
@@ -34,10 +34,47 @@
     "contributes": {
         "commands": [
             {
-                "command": "excalidraw.theme.update",
-                "title": "Color Theme",
-                "icon": "$(color-mode)",
-                "category": "Excalidraw"
+                "command": "excalidraw.updateTheme",
+                "title": "Choose Excalidraw Theme",
+                "icon": "$(color-mode)"
+            },
+            {
+                "command": "excalidraw.toggleTheme",
+                "title": "Toggle Excalidraw Theme",
+                "icon": "$(color-mode)"
+            },
+            {
+                "command": "excalidraw.showSource",
+                "title": "Show Source",
+                "icon": "$(go-to-file)"
+            },
+            {
+                "command": "excalidraw.showEditor",
+                "title": "Open in Excalidraw",
+                "icon": {
+                    "dark": "./assets/menu-icon-dark.png",
+                    "light": "./assets/menu-icon-light.png"
+                }
+            },
+            {
+                "command": "excalidraw.showImage",
+                "title": "Show Image",
+                "icon": "$(go-to-file)"
+            },
+            {
+                "command": "excalidraw.showSourceToSide",
+                "title": "Show Source to the Side",
+                "icon": "$(go-to-file)"
+            },
+            {
+                "command": "excalidraw.showEditorToSide",
+                "title": "Open in Excalidraw to the Side",
+                "icon": "./assets/logo.png"
+            },
+            {
+                "command": "excalidraw.showImageToSide",
+                "title": "Show Image to the Side",
+                "icon": "$(go-to-file)"
             }
         ],
         "configuration": [
@@ -70,11 +107,64 @@
             }
         },
         "menus": {
+            "commandPalette": [
+                {
+                    "command": "excalidraw.updateTheme",
+                    "when": "activeCustomEditorId == 'editor.excalidraw'"
+                },
+                {
+                    "command": "excalidraw.toggleTheme",
+                    "when": "false"
+                },
+                {
+                    "command": "excalidraw.showSource",
+                    "when": "false"
+                },
+                {
+                    "command": "excalidraw.showEditor",
+                    "when": "false"
+                },
+                {
+                    "command": "excalidraw.showImage",
+                    "when": "false"
+                },
+                {
+                    "command": "excalidraw.showSourceToSide",
+                    "when": "false"
+                },
+                {
+                    "command": "excalidraw.showEditorToSide",
+                    "when": "false"
+                },
+                {
+                    "command": "excalidraw.showImageToSide",
+                    "when": "false"
+                }
+            ],
             "editor/title": [
                 {
-                    "command": "excalidraw.theme.update",
+                    "command": "excalidraw.showEditor",
+                    "alt": "excalidraw.showEditorToSide",
+                    "when": "activeCustomEditorId != 'editor.excalidraw' && resourceFilename =~ /^.+\\.excalidraw(.png)?$/",
+                    "group": "navigation"
+                },
+                {
+                    "command": "excalidraw.showSource",
+                    "alt": "excalidraw.showSourceToSide",
                     "group": "navigation",
-                    "when": "resourceFilename =~ /^.+\\.excalidraw$/ || resourceFilename =~ /^.+\\.excalidraw\\.svg$/ || resourceFilename =~ /^.+\\.excalidraw\\.png$/"
+                    "when": "activeCustomEditorId == 'editor.excalidraw' && resourceFilename =~ /^.+\\.excalidraw(.svg)?$/"
+                },
+                {
+                    "command": "excalidraw.showImage",
+                    "alt": "excalidraw.showImageToSide",
+                    "when": "activeCustomEditorId == 'editor.excalidraw' && resourceFilename =~ /^.+\\.excalidraw.png$/",
+                    "group": "navigation"
+                },
+                {
+                    "command": "excalidraw.toggleTheme",
+                    "alt": "excalidraw.updateTheme",
+                    "when": "activeCustomEditorId == 'editor.excalidraw' && resourceFilename =~ /^.+\\.excalidraw(.svg|.png)?$/",
+                    "group": "navigation"
                 }
             ]
         },
