+++ extension/src/document.ts
