+++ extension/src/editor.ts
@@ -2,7 +2,7 @@ import * as vscode from "vscode";
 import * as path from "path";
 import { Base64 } from "js-base64";
 
-import { ExcalidrawDocument } from "./ExcalidrawDocument";
+import { ExcalidrawDocument } from "./document";
 
 export class ExcalidrawEditorProvider
   implements vscode.CustomEditorProvider<ExcalidrawDocument>
