+++ extension/src/uri-handler.ts
@@ -1,0 +1,24 @@
+import * as vscode from "vscode";
+import { ExcalidrawEditor } from "./editor";
+
+export class ExcalidrawUriHandler implements vscode.UriHandler {
+  public static register() {
+    const provider = new ExcalidrawUriHandler();
+    const providerRegistration = vscode.window.registerUriHandler(provider);
+    return providerRegistration;
+  }
+
+  public async handleUri(uri: vscode.Uri) {
+    try {
+      const hash = new URLSearchParams(uri.fragment);
+      const libraryUrl = hash.get("addLibrary");
+      const csrfToken = hash.get("token");
+      if (libraryUrl && csrfToken) {
+        ExcalidrawEditor.importLibrary(libraryUrl, csrfToken);
+        vscode.window.showInformationMessage("Library added successfully!");
+      }
+    } catch (e) {
+      console.error(e);
+    }
+  }
+}
